initial parameter organisation move ipaddress parrameter and add snat selection option to network offering
