+++ api/src/main/java/com/cloud/offering/NetworkOffering.java
@@ -151,4 +151,6 @@ public interface NetworkOffering extends InfrastructureEntity, InternalIdentity,
     String getServicePackage();
 
     Date getCreated();
+
+    boolean isSelectSnatIpAllowed();
 }
