+++ api/src/main/java/org/apache/cloudstack/api/command/admin/network/CreateNetworkCmdByAdmin.java
@@ -16,7 +16,6 @@
 // under the License.
 package org.apache.cloudstack.api.command.admin.network;
 
-import org.apache.cloudstack.api.ApiArgValidator;
 import org.apache.log4j.Logger;
 
 import org.apache.cloudstack.api.APICommand;
@@ -43,14 +42,6 @@ public class CreateNetworkCmdByAdmin extends CreateNetworkCmd implements AdminCm
     @Parameter(name=ApiConstants.HIDE_IP_ADDRESS_USAGE, type=CommandType.BOOLEAN, description="when true ip address usage for the network will not be exported by the listUsageRecords API")
     private Boolean hideIpAddressUsage;
 
-    @Parameter(name = ApiConstants.ROUTER_IP, type = CommandType.STRING, description = "IPV4 address to be assigned to a router in a shared network", since = "4.16",
-            validations = {ApiArgValidator.NotNullOrEmpty})
-    private String routerIp;
-
-    @Parameter(name = ApiConstants.ROUTER_IPV6, type = CommandType.STRING, description = "IPV6 address to be assigned to a router in a shared network", since = "4.16",
-            validations = {ApiArgValidator.NotNullOrEmpty})
-    private String routerIpv6;
-
     /////////////////////////////////////////////////////
     /////////////////// Accessors ///////////////////////
     /////////////////////////////////////////////////////
@@ -73,11 +64,4 @@ public class CreateNetworkCmdByAdmin extends CreateNetworkCmd implements AdminCm
         return false;
     }
 
-    public String getRouterIp() {
-        return routerIp;
-    }
-
-    public String getRouterIpv6() {
-        return routerIpv6;
-    }
 }
